requirements/testing.txt
pytest==3.7.4
pytest-cov==2.6.0
pytest-django==3.4.2
